+++ src/utils/get-web3.js
@@ -1,6 +1,6 @@
 import Web3 from 'web3';
 
-const getWeb3 = new Promise((resolve) => {
+const getWeb3 = new Promise((resolve, reject) => {
   // wait for loading completion before loading web3, to be sure it's
   // already injected
   window.addEventListener('load', () => {
@@ -11,11 +11,10 @@ const getWeb3 = new Promise((resolve) => {
       // Use MetaMask's provider.
       web3js = new Web3(window.web3.currentProvider);
       results = { web3: web3js };
-      console.log('Injected web3 detected.');
       resolve(results);
     } else {
       // user is not running MetaMask?
-      console.log('Is MetaMask running?');
+      reject(new Error('Is MetaMask running?'));
     }
   });
 });